Refactor functions in data generation script for easier re-usability
