+++ scripts/generate_data.py
@@ -5,166 +5,10 @@ import click
 import geopandas as gpd
 import pandas as pd
 from loguru import logger
-from rasterstats import zonal_stats
-from shapely import wkt
-from tqdm.auto import tqdm
 
 from src.config import settings
-from src.data_processing import aod, era5, gee_utils, geom_utils, ndvi
-
-
-def generate_locations_with_dates_df(
-    df, start_date, end_date, id_col="id", date_col="date"
-):
-    # We create a dummy date column just so we can use the ffill technique to construct one row for each date per station.
-    df = df.copy()
-    df[date_col] = pd.to_datetime(start_date, format="%Y-%m-%d")
-    df = (
-        df.groupby([id_col])
-        .apply(
-            lambda x: x.set_index(date_col)
-            .reindex(pd.date_range(start=start_date, end=end_date))
-            .ffill()
-            .rename_axis(date_col)
-            .reset_index()
-        )
-        .droplevel(id_col)
-    )
-    df[date_col] = df[date_col].dt.date
-    return df
-
-
-def collect_gee_datasets(gee_datasets, start_date, end_date, locations_df, id_col="id"):
-    gee_dfs = {}
-    for gee_index, gee_dataset in enumerate(gee_datasets):
-
-        logger.info(
-            f"Collecting GEE data ({gee_index+1} / {len(gee_datasets)}): {gee_dataset}"
-        )
-
-        collection_id = gee_dataset["collection_id"]
-        bands = gee_dataset["bands"]
-        preprocessors = gee_dataset["preprocessors"]
-
-        # For recording all dfs before concatenating later on
-        all_dfs = []
-
-        # Iterate through stations
-        for index, location in tqdm(locations_df.iterrows(), total=len(locations_df)):
-            # Generate station data
-            station_gee_values_df = gee_utils.generate_aoi_tile_data(
-                collection_id,
-                start_date,
-                end_date,
-                location.latitude,
-                location.longitude,
-                bands=bands,
-                cloud_filter=False,
-            )
-            # Set the ID so we can join back the data later on
-            station_gee_values_df[id_col] = location[id_col]
-
-            # Pre-process
-            for preprocessor, params in preprocessors:
-                station_gee_values_df = preprocessor(station_gee_values_df, params)
-
-            # Add to main df
-            all_dfs.append(station_gee_values_df)
-
-        gee_dfs[collection_id] = pd.concat(all_dfs, axis=0, ignore_index=True)
-
-    return gee_dfs
-
-
-def join_datasets(
-    locations_df,
-    start_date,
-    end_date,
-    gee_dfs,
-    hrsl_df,
-    id_col,
-    date_col="date",
-    ground_truth_df=None,
-    admin_bounds_gdf=None,
-):
-    # Create DF with locations + start_date, end_date
-    base_df = generate_locations_with_dates_df(
-        locations_df, start_date, end_date, id_col=id_col, date_col=date_col
-    )
-
-    # Merge HRSL
-    # HRSL is a slow-moving feature, and so does not change depending on the date.
-    base_df = base_df.merge(hrsl_df, on=[id_col], how="left")
-
-    # Merge GEE dfs
-    for _, gee_df in gee_dfs.items():
-        base_df = base_df.merge(gee_df, on=[id_col, date_col], how="left")
-
-    # Ground Truth
-    if ground_truth_df is not None:
-        base_df = base_df.merge(ground_truth_df, on=[id_col, date_col], how="left")
-
-    # Admin Bounds
-    if admin_bounds_gdf is not None:
-        base_df = gpd.GeoDataFrame(
-            base_df,
-            geometry=gpd.points_from_xy(base_df["longitude"], base_df["latitude"]),
-            crs="EPSG:4326",
-        )
-        # First, we want to retain only the unique columns in the right GDF to avoid duplicate columns (and consequent column renames).
-        cols_unique_to_adm_bounds = admin_bounds_gdf.columns.difference(
-            base_df.columns
-        ).tolist() + ["geometry"]
-        admin_bounds_gdf = admin_bounds_gdf[cols_unique_to_adm_bounds]
-
-        # Perform the actual spatial join
-        base_df = gpd.sjoin(base_df, admin_bounds_gdf, predicate="within", how="left")
-
-        # Drop unnecessary columns generated by the sjoin
-        base_df.drop(["index_right"], axis=1, inplace=True)
-
-    # Sorting
-    base_df = base_df.sort_values(by=[id_col, date_col])
-
-    return base_df
-
-
-def load_ground_truth(csv_path):
-    df = pd.read_csv(csv_path)
-    df["date"] = pd.to_datetime(df["date"]).dt.date
-    return df
-
-
-def generate_bboxes(
-    locations_df,
-    bbox_size_km,
-    geometry_col="geometry",
-):
-    locations_df = locations_df.copy()
-    locations_df[geometry_col] = locations_df.apply(
-        lambda row: geom_utils.generate_bbox_wkt(
-            row["latitude"], row["longitude"], distance_km=bbox_size_km
-        ),
-        axis=1,
-    )
-    return locations_df
-
-
-def collect_hrsl(locations_df, hrsl_tif, id_col, bbox_size_km):
-    locations_df = generate_bboxes(locations_df, bbox_size_km=bbox_size_km)
-    # locations_df.to_csv(settings.DATA_DIR / "debug_locations.csv")
-    hrsl_gdf = gpd.GeoDataFrame(
-        locations_df, geometry=locations_df["geometry"].apply(wkt.loads)
-    )
-    hrsl_gdf["total_population"] = pd.DataFrame(
-        zonal_stats(vectors=hrsl_gdf["geometry"], raster=hrsl_tif, stats="sum")
-    )["sum"]
-
-    # Retain only id_col and total_population to save on space.
-    hrsl_df = pd.DataFrame(hrsl_gdf.drop(columns="geometry"))[
-        [id_col, "total_population"]
-    ]
-    return hrsl_df
+from src.data_processing import general_utils, hrsl
+from src.data_processing.gee import aod, era5, gee_utils, ndvi
 
 
 @click.command()
@@ -233,7 +77,8 @@ def main(
 
     # Read in ground truth if any
     if ground_truth_csv:
-        ground_truth_df = load_ground_truth(ground_truth_csv)
+        ground_truth_df = pd.read_csv(ground_truth_csv)
+        ground_truth_df["date"] = pd.to_datetime(ground_truth_df["date"]).dt.date
         logger.info(f"Generating dataset with ground truth from {ground_truth_csv}")
     else:
         ground_truth_df = None
@@ -249,7 +94,7 @@ def main(
 
     # Compute HRSL stats
     logger.info("Computing population sums...")
-    hrsl_df = collect_hrsl(
+    hrsl_df = hrsl.collect_hrsl(
         locations_df, hrsl_tif, id_col=id_col, bbox_size_km=BBOX_SIZE_KM
     )
 
@@ -302,7 +147,7 @@ def main(
         },
     ]
 
-    gee_dfs = collect_gee_datasets(
+    gee_dfs = general_utils.collect_gee_datasets(
         gee_datasets, start_date, end_date, locations_df, id_col=id_col
     )
 
@@ -321,7 +166,7 @@ def main(
 
     # Generate final DF and save to CSV
     logger.info("Joining all datasets together...")
-    base_df = join_datasets(
+    base_df = general_utils.join_datasets(
         locations_df,
         start_date,
         end_date,
